Refactor view code into modal.js
